Disable_people_DjangoApis/Django/requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bidict==0.22.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-extra-fields==3.4.1
EasyROI==1.0.3
executing==1.2.0
filelock==3.7.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h11==0.14.0
hydra-core==1.3.1
idna==3.4
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
numpy==1.24.2
oauthlib==3.2.2
oldest-supported-numpy==2022.11.19
omegaconf==2.3.0
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==1.14.0
shapely==2.0.1
simple-websocket==0.9.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
ultralytics==8.0.35
urllib3==1.26.14
virtualenv==20.15.0
wcwidth==0.2.6
Werkzeug==2.2.2
wsproto==1.2.0
django-cors-headers==3.14.0
django-jalali==6.0.1
drf_extra_fields==3.4.1
Pillow==9.4.0
auth-token==0.0.3
django-rest-authtoken==2.1.4